+++ spec/test_helper.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-ENV["RAILS_ENV"] = "test"
-
-require "bundler"
-Bundler.setup
-require "active_support/test_case"
-require "active_support/testing/autorun"